Moved storage-related methods from LocalEntity to LocalStorage (static) + method renamed
